update loan with lending pool

+++ backend/src/modules/offers/offers.service.ts
@@ -105,31 +105,46 @@ export class OffersService implements OnModuleInit {
           case 'LoanStarted': {
             const loanId = event.args.loanId;
             const offer = await this.findById(loanId);
-            await Promise.all([
-              this.offer.update(loanId, { status: OfferStatus.FILLED }),
-              this.offer.deleteAllAccept(loanId),
-              this.order.update(offer.order, { status: OrderStatus.FILLED }),
-            ]);
+            if (offer) {
+              await Promise.all([
+                this.offer.update(loanId, { status: OfferStatus.FILLED }),
+                this.offer.deleteAllAccept(loanId),
+                this.order.update(offer.order, { status: OrderStatus.FILLED }),
+              ]);
+            } else {
+              await this.order.update(loanId, { status: OrderStatus.FILLED });
+            }
+
             break;
           }
           case 'LoanRepaid': {
             const loanId = event.args.loanId;
             const offer = await this.findById(loanId);
-            await Promise.all([
-              this.offer.update(loanId, { status: OfferStatus.REPAID }),
-              this.order.update(offer.order, { status: OrderStatus.REPAID }),
-            ]);
+            if (offer) {
+              await Promise.all([
+                this.offer.update(loanId, { status: OfferStatus.REPAID }),
+                this.order.update(offer.order, { status: OrderStatus.REPAID }),
+              ]);
+            } else {
+              await this.order.update(loanId, { status: OrderStatus.REPAID });
+            }
+
             break;
           }
           case 'LoanLiquidated': {
             const loanId = event.args.loanId;
             const offer = await this.findById(loanId);
-            await Promise.all([
-              this.offer.update(loanId, { status: OfferStatus.LIQUIDATED }),
-              this.order.update(offer.order, {
-                status: OrderStatus.LIQUIDATED,
-              }),
-            ]);
+            if (offer) {
+              await Promise.all([
+                this.offer.update(loanId, { status: OfferStatus.LIQUIDATED }),
+                this.order.update(offer.order, {
+                  status: OrderStatus.LIQUIDATED,
+                }),
+              ]);
+            } else {
+              await this.order.update(loanId, { status: OrderStatus.REPAID });
+            }
+
             break;
           }
           default: {
